+++ javascript/node/selenium-webdriver/CHANGES.md
@@ -5,9 +5,12 @@
     `webdriver.promise.consume`. Requires harmony support (run with
     `node --harmony-generators` in `v0.11.x`).
 * Promise A+ compliance: a promise may no longer resolve to itself.
-* For consistency with other language bindings, deprecated
-    `UnhandledAlertError#getAlert` and added `#getAlertText`.
-    `getAlert` will be removed in `2.45.0`.
+* Changes for consistency with other language bindings:
+    * Added the `chrome.ChromeDriver` class, which may be used to create a new
+        ChromeDriver session (vs. using `webdriver.Builder`).
+    * Added the `phantomjs.PhantomDriverJs` class.
+    * Deprecated `UnhandledAlertError#getAlert` and added `#getAlertText`.
+        `getAlert` will be removed in `2.45.0`.
 * FIXED: 7563: Mocha integration no longer disables timeouts. Default Mocha
     timeouts apply (2000 ms) and may be changed using `this.timeout(ms)`.